LPS-76957 Move to web
