+++ modules/apps/social-networking/social-networking-web/src/main/java/com/liferay/social/networking/web/internal/members/social/MembersActivityInterpreter.java
@@ -31,7 +31,6 @@ import com.liferay.social.kernel.model.BaseSocialActivityInterpreter;
 import com.liferay.social.kernel.model.SocialActivity;
 import com.liferay.social.kernel.model.SocialActivityInterpreter;
 import com.liferay.social.networking.constants.SocialNetworkingPortletKeys;
-import com.liferay.social.networking.members.social.MembersActivityKeys;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
