+++ modules/apps/social-networking/social-networking-web/src/main/java/com/liferay/social/networking/web/internal/members/social/MembersRequestInterpreter.java
@@ -31,8 +31,6 @@ import com.liferay.social.kernel.model.SocialRequestFeedEntry;
 import com.liferay.social.kernel.model.SocialRequestInterpreter;
 import com.liferay.social.kernel.service.SocialActivityLocalService;
 import com.liferay.social.networking.constants.SocialNetworkingPortletKeys;
-import com.liferay.social.networking.members.social.MembersActivityKeys;
-import com.liferay.social.networking.members.social.MembersRequestKeys;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
