+++ modules/apps/social-networking/social-networking-web/src/main/java/com/liferay/social/networking/web/internal/social/BaseSocialNetworkingActivityInterpreter.java
@@ -23,7 +23,7 @@ import com.liferay.portal.kernel.util.ResourceBundleLoaderUtil;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.social.kernel.model.BaseSocialActivityInterpreter;
 import com.liferay.social.kernel.model.SocialActivity;
-import com.liferay.social.networking.friends.social.FriendsActivityKeys;
+import com.liferay.social.networking.web.internal.friends.social.FriendsActivityKeys;
 
 import org.osgi.service.component.annotations.Reference;
