+++ modules/apps/social-networking/social-networking-web/src/main/java/com/liferay/social/networking/web/internal/social/BaseSocialNetworkingRequestInterpreter.java
@@ -32,8 +32,8 @@ import com.liferay.social.kernel.model.SocialRequest;
 import com.liferay.social.kernel.model.SocialRequestFeedEntry;
 import com.liferay.social.kernel.service.SocialActivityLocalService;
 import com.liferay.social.kernel.service.SocialRelationLocalService;
-import com.liferay.social.networking.friends.social.FriendsActivityKeys;
-import com.liferay.social.networking.friends.social.FriendsRequestKeys;
+import com.liferay.social.networking.web.internal.friends.social.FriendsActivityKeys;
+import com.liferay.social.networking.web.internal.friends.social.FriendsRequestKeys;
 
 /**
  * @author Adolfo Pérez
