+++ modules/apps/social-networking/social-networking-web/src/main/java/com/liferay/social/networking/web/internal/util/WallUtil.java
@@ -24,8 +24,8 @@ import com.liferay.portal.kernel.util.HtmlUtil;
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.Validator;
-import com.liferay.social.networking.configuration.SocialNetworkingServiceConfigurationValues;
 import com.liferay.social.networking.constants.SocialNetworkingPortletKeys;
+import com.liferay.social.networking.web.internal.configuration.SocialNetworkingServiceConfigurationValues;
 
 /**
  * @author Zsolt Berentey
